frontend/src/component/upload/DropZone/useDragAndDrop.ts
import {useState} from 'react'

const useDragAndDrop = () => {
  const [dragOver, setDragOver] = useState(false)
  const [droppedFile, setDroppedFile] = useState<File | null>(null)

  const handleDrop = (e: React.DragEvent<HTMLDivElement>) => {
    e.preventDefault()
    if (e.dataTransfer.items) {
      const file = e.dataTransfer.items[0].getAsFile()
      if (file) {
        setDroppedFile(file)
      }
    }
    setDragOver(false)
  }

  const handleDragOver = (e: React.DragEvent<HTMLDivElement>) => {
    e.preventDefault()
    setDragOver(true)
  }

  const handleDragLeave = (e: React.DragEvent<HTMLDivElement>) => {
    e.preventDefault()
    setDragOver(false)
  }

  return {dragOver, droppedFile, handleDrop, handleDragOver, handleDragLeave}
}

export {useDragAndDrop}
